climbing: move layers to folder
